working on step 4
